Break out separate component for attributes
